refactor: merge executor and runner together as the functionalities are related
